chore(DB-12): fixed nestjs api import issue from separate moules
