+++ apps-api/resource-hub-api/src/app.module.ts
@@ -1,10 +1,10 @@
 import { GraphQLModule } from '@nestjs/graphql';
 import { MiddlewareConsumer, Module } from '@nestjs/common';
-import { UserPublicModule } from '@molitio/resource-hub-modules';
 import { ApolloDriver, ApolloDriverConfig } from '@nestjs/apollo';
 import { AppService } from './app.service.js';
 import { AppController } from './app.controller.js';
 import { LoggerMiddleware } from './logger/logger.middleware.js';
+import { UserPublicModule } from './modules';
 
 @Module({
     imports: [
