+++ apps-api/resource-hub-api/src/modules/UserPublic/UserPublic.mutation.ts
@@ -1,5 +1,5 @@
 import { Args, Mutation, Resolver } from '@nestjs/graphql';
-import { UserPublic } from './UserPublic.type';
+import { UserPublic } from './UserPublic.type.js';
 
 @Resolver((of: any) => UserPublic)
 export class UserPublicResolverMutation {
