+++ apps-api/resource-hub-api/src/modules/UserPublic/UserPublic.resolver.ts
@@ -1,6 +1,6 @@
 import { Resolver } from '@nestjs/graphql';
-import { UserPublicService } from './UserPublic.service';
-import { UserPublic } from './UserPublic.type';
+import { UserPublicService } from './UserPublic.service.js';
+import { UserPublic } from './UserPublic.type.js';
 
 @Resolver((of: any) => UserPublic)
 export class UserPublicResolver {
