+++ apps-api/resource-hub-api/src/modules/UserPublic/UserPublic.type.ts
