+++ apps-api/resource-hub-api/src/modules/UserPublic/index.ts
@@ -1,0 +1,5 @@
+export * from './UserPublic.resolver.js';
+export * from './UserPublic.query.js';
+export * from './UserPublic.mutation.js';
+export * from './UserPublic.module.js';
+export * from './UserPublic.type.js';
