--- packages-api/resource-hub-modules/README.md
@@ -1,7 +1,0 @@
-# resource-hub-modules
-
-Resource Hub Modules is a repository designated to be a single source of truth of molitio schema models. It provides the means to generate type safe data contracts to be consumed by the ui and the back and as well.
-
-
-       
-
