--- packages-api/resource-hub-modules/package.json
@@ -1,45 +1,0 @@
-{
-    "name": "@molitio/resource-hub-modules",
-    "version": "0.1.1",
-    "description": "Isolated Resource Hub Modules",
-    "main": "src/index.ts",
-    "type": "module",
-    "scripts": {
-        "build": "rm -rf ./dist && tsc --build --verbose",
-        "dev": "yarn tsc --watch"
-    },
-    "devDependencies": {
-        "@graphql-codegen/add": "3.2.3",
-        "@graphql-codegen/cli": "5.0.2",
-        "@graphql-codegen/client-preset": "1.2.3",
-        "@graphql-codegen/graphql-modules-preset": "2.5.1",
-        "@graphql-codegen/near-operation-file-preset": "2.4.4",
-        "@graphql-codegen/schema-ast": "2.6.0",
-        "@graphql-codegen/typescript": "2.7.3",
-        "@graphql-codegen/typescript-graphql-request": "4.5.8",
-        "@graphql-codegen/typescript-operations": "2.5.10",
-        "@graphql-codegen/typescript-react-apollo": "3.3.7",
-        "@graphql-codegen/typescript-react-query": "4.0.6",
-        "@graphql-codegen/typescript-resolvers": "2.7.3",
-        "@graphql-typed-document-node/core": "3.1.1",
-        "@types/node": "16.11.26",
-        "class-validator": "0.14.1",
-        "core-js": "3.30.2",
-        "graphql": "16.8.1",
-        "graphql-request": "5.1.0",
-        "graphql-scalars": "1.23.0",
-        "graphql-tag": "2.12.6",
-        "reflect-metadata": "0.2.1",
-        "tsc": "2.0.4",
-        "tslib": "2.6.2",
-        "typescript": "5.2.2"
-    },
-    "stableVersion": "0.1.1",
-    "dependencies": {
-        "@nestjs/common": "^10.0.0",
-        "@nestjs/core": "10.3.7",
-        "@nestjs/graphql": "12.1.1",
-        "rxjs": "7.8.1",
-        "ts-node": "10.9.2"
-    }
-}
