--- packages-api/resource-hub-modules/src/UserPublic/index.ts
@@ -1,5 +1,0 @@
-export * from './UserPublic.resolver';
-export * from './UserPublic.query';
-export * from './UserPublic.mutation';
-export * from './UserPublic.module';
-export * from './UserPublic.type';
