--- packages-api/resource-hub-modules/tsconfig.json
@@ -1,14 +1,0 @@
-{
-    "compilerOptions": {
-        "outDir": "./dist",
-        "baseUrl": "./src",
-        "target": "es2022",
-        "moduleResolution": "Bundler",
-        "declaration": true,
-        "skipLibCheck": true,
-        "declarationMap": false,
-        "emitDecoratorMetadata": true,
-        "experimentalDecorators": true
-    },
-    "include": ["src/**/*.ts"]
-}
