--- packages-api/resource-hub-modules/typedoc.json
@@ -1,4 +1,0 @@
-{
-    "$schema": "https://typedoc.org/schema.json",
-    "extends": ["../../typedoc.base.json"]
-}
